Fixed reload bug

+++ app/src/main/java/com/guminegor/example/downloadingpicture/Settings.java
@@ -3,8 +3,10 @@ package com.guminegor.example.downloadingpicture;
 import android.content.SharedPreferences;
 import android.os.Bundle;
 import android.os.Environment;
+import android.support.v4.app.NavUtils;
 import android.support.v7.app.AppCompatActivity;
 import android.support.v7.widget.Toolbar;
+import android.view.MenuItem;
 import android.view.View;
 import android.widget.Button;
 import android.widget.RadioButton;
@@ -81,41 +83,50 @@ public class Settings extends AppCompatActivity implements View.OnClickListener{
                 break;
 
             case R.id.settings_btn_save:
-                cachePath = Environment.getExternalStorageDirectory() + "/" + images[currentImage];
-                cachedImage = new File(cachePath);
-                boolean deleted = cachedImage.delete();
-                ed.putInt("currentImage", newImage);
-                ed.putInt("downloadStatus", 0);
-                ed.commit();
-
-//                if (prefs.getInt("downloadStatus", 0) != 1){
-//
-//                }
-//                else{
-//                    Toast.makeText(Settings.this, R.string.warning_file_downloading, Toast.LENGTH_LONG).show();
-//                }
+                if (prefs.getInt("downloadStatus", 0) != 1){
+                    cachePath = Environment.getExternalStorageDirectory() + "/" + images[currentImage];
+                    cachedImage = new File(cachePath);
+                    boolean deleted = cachedImage.delete();
+                    ed.putInt("currentImage", newImage);
+                    ed.putInt("downloadStatus", 0);
+                    ed.commit();
+                    Toast.makeText(Settings.this, R.string.saved, Toast.LENGTH_LONG).show();
+                }
+                else{
+                    Toast.makeText(Settings.this, R.string.warning_file_downloading, Toast.LENGTH_LONG).show();
+                }
 
                 break;
 
             case R.id.settings_btn_clear:
-                Toast.makeText(Settings.this, R.string.warning_file_downloading, Toast.LENGTH_LONG).show();
-                cachePath = Environment.getExternalStorageDirectory() + "/" + images[currentImage];
-                cachedImage = new File(cachePath);
-                boolean del = cachedImage.delete();
-                ed.putInt("downloadStatus", 0);
-                ed.commit();
-
-//                if (prefs.getInt("downloadStatus", 0) != 1){
-//
-//                }
-//                else{
-//                    Toast.makeText(Settings.this, R.string.warning_file_downloading, Toast.LENGTH_LONG).show();
-//                }
+                if (prefs.getInt("downloadStatus", 0) != 1){
+                    cachePath = Environment.getExternalStorageDirectory() + "/" + images[currentImage];
+                    cachedImage = new File(cachePath);
+                    boolean del = cachedImage.delete();
+                    ed.putInt("downloadStatus", 0);
+                    ed.commit();
+                    Toast.makeText(Settings.this, R.string.deleted, Toast.LENGTH_LONG).show();
+                }
+                else{
+                    Toast.makeText(Settings.this, R.string.warning_file_downloading, Toast.LENGTH_LONG).show();
+                }
                break;
 
             default:
                 break;
         }
     }
+
+    @Override
+    public boolean onOptionsItemSelected(MenuItem item) {
+        switch (item.getItemId()) {
+            case android.R.id.home:
+                NavUtils.navigateUpFromSameTask(this);
+                return true;
+            default:
+                return super.onOptionsItemSelected(item);
+        }
+    }
+
 
 }
